+++ src/odata/serialization/function-serializer.ts
@@ -37,7 +37,6 @@ const serializers: { [TContainer in keyof typeof functionContainers]: ContainerS
   collections: {
     in: (v1, v2) => `${v1} in ${v2}`,
     distinct: null,
-    filterDefined: null,
     first: null
   },
   comparison: {
@@ -50,6 +49,7 @@ const serializers: { [TContainer in keyof typeof functionContainers]: ContainerS
     lowerOrEqual: (v1, v2) => `${v1} le ${v2}`
   },
   internal: {
+    ifUndefined: null,
     mergeObjects: null
   },
   mathematics: {